+++ .env.example
@@ -1,5 +1,5 @@
-D5DK8S_CONFIG_BOT_TOKEN="secretToken"
-D5DK8S_CONFIG_DATABASE_URL="postgresql+asyncpg://postgres:123456@postgres:5432/kube_inspector_bot"
+BOT_CONFIG_BOT_TOKEN="secretToken"
+BOT_CONFIG_DATABASE_URL="postgresql+asyncpg://postgres:123456@postgres:5432/kube_inspector_bot"
 
 # for the postgres image in docker-compose.yml
 POSTGRES_PASSWORD=123456
