+++ Dockerfile
@@ -1,7 +1,7 @@
 FROM python:3.11-slim-bullseye
-WORKDIR /d5dk8s
+WORKDIR /bot
 
-COPY d5dk8s .
+COPY bot .
 COPY requirements.txt .
 COPY alembic.ini /
 
@@ -9,4 +9,4 @@ RUN ["pip", "install", "-r", "requirements.txt"]
 
 WORKDIR /
 
-ENTRYPOINT ["python", "-m", "d5dk8s"]
+ENTRYPOINT ["python", "-m", "bot"]
