+++ bot.example.yml
