+++ bot/database/database.py
@@ -1,4 +1,4 @@
 from sqlalchemy.ext.asyncio import create_async_engine
-from d5dk8s.config import Config
+from bot.config import Config
 
 engine = create_async_engine(Config.get('database.url'), echo=False)
