+++ bot/database/tables.py
